Training_autoML_pipeline/custom-data-entities-image/requirements.txt
pandas==2.0.0
numpy==1.25.2
google-api-core
google-api-python-client
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.41.0
google-cloud-artifact-registry==1.10.0
google-cloud-bigquery==3.17.2
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.4.1
google-cloud-datastore==1.15.5
google-cloud-jupyter-config==0.0.5
google-cloud-language==2.12.0
google-cloud-monitoring==2.18.0
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
tqdm
scikit-learn
db-dtypes
nltk